Abstract away function / gradient calls in algorithms
